+++ src/renderer.h
@@ -14,13 +14,13 @@
 class renderer
 {
 public:
-    renderer(const glm::vec2& screen_size, Ymir::ResourceManager<Ymir::Program, Ymir::ProgramResourceLoader>& programManager, Ymir::ResourceManager<Ymir::Mesh>& meshManager);
+    renderer(const glm::ivec2& screen_size, Ymir::ResourceManager<Ymir::Program, Ymir::ProgramResourceLoader>& programManager, Ymir::ResourceManager<Ymir::Mesh>& meshManager);
     void boot();
     void shutdown();
     
     bool step(std::shared_ptr<game_object> world);
 private:
-    const glm::vec2 SCREEN_SIZE;
+    const glm::ivec2 SCREEN_SIZE;
     Ymir::ResourceManager<Ymir::Program, Ymir::ProgramResourceLoader>& _programManager;
     Ymir::ResourceManager<Ymir::Mesh>& _meshManager;
     std::unique_ptr<deferred_render_visitor> _deferred_renderer;
